+++ dev-tools/omicron-dev/src/lib.rs
@@ -1,0 +1,19 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+//! Developer tool for easily running bits of Omicron.
+
+#[cfg(feature = "include-cert")]
+mod cert;
+#[cfg(feature = "include-clickhouse")]
+mod clickhouse;
+#[cfg(feature = "include-db")]
+mod db;
+mod dispatch;
+#[cfg(feature = "include-mgs")]
+mod mgs;
+#[cfg(feature = "include-nexus")]
+mod nexus;
+
+pub use dispatch::*;
