+++ dev-tools/xtask/src/main.rs
@@ -16,6 +16,7 @@ mod clippy;
 mod download;
 #[cfg_attr(not(target_os = "illumos"), allow(dead_code))]
 mod external;
+mod omicron_dev;
 mod usdt;
 
 #[cfg(target_os = "illumos")]
@@ -53,6 +54,9 @@ enum Cmds {
     ///
     /// For more information, see dev-tools/openapi-manager/README.adoc.
     Openapi(external::External),
+
+    /// Run Omicron development tasks
+    OmicronDev(omicron_dev::OmicronDevArgs),
 
     #[cfg(target_os = "illumos")]
     /// Build a TUF repo
@@ -98,6 +102,7 @@ async fn main() -> Result<()> {
         Cmds::CheckWorkspaceDeps => check_workspace_deps::run_cmd(),
         Cmds::Download(args) => download::run_cmd(args).await,
         Cmds::Openapi(external) => external.exec_bin("openapi-manager"),
+        Cmds::OmicronDev(args) => omicron_dev::run_cmd(args).await,
         #[cfg(target_os = "illumos")]
         Cmds::Releng(external) => {
             external.cargo_args(["--release"]).exec_bin("omicron-releng")
